Add repository for tasks (with a naive "DB" implementation)
The "DB" is just something to store the items in (a memory like
singleton storage). It's not tested as in a production-like env
it would be a Mongo/Dynamo or any other NOSQL type DB for document
storing.

The `abstract` module is a generic storage type implementation that
can take any model and use it.